backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
vobject>=0.9.6
pyyaml>=6.0
